Rename sponge to ternary
