+++ Cargo.toml
@@ -31,9 +31,9 @@ x25519 = [ "x25519-dalek" ]
 random = [ "getrandom" ]
 aes = [ "aes-gcm", "cipher" ]
 blake2b = [ "blake2", "digest" ]
-sponge = []
-curl-p = [ "sponge", "bee-ternary" ]
-kerl_deprecated_do_not_use = [ "sponge", "bee-ternary", "byteorder", "lazy_static", "tiny-keccak" ]
+ternary_hashes = []
+curl-p = [ "ternary_hashes", "bee-ternary" ]
+kerl_deprecated_do_not_use = [ "ternary_hashes", "bee-ternary", "byteorder", "lazy_static", "tiny-keccak" ]
 ternary_keys = []
 ternary_signatures = []
 wots_deprecated_do_not_use = [ "ternary_keys", "ternary_signatures", "bee-ternary", "sha3", "zeroize", "rand", "bee-common-derive" ]
