+++ src/hashes/ternary/curl_p.rs
