+++ src/hashes/ternary/kerl/bigint/error.rs
