+++ src/hashes/ternary/kerl/bigint/macros.rs
