+++ src/hashes/ternary/kerl/bigint/sealed.rs
