+++ src/hashes/ternary/kerl/bigint/split_integer.rs
