+++ src/hashes/ternary/kerl/bigint/t243/mod.rs
@@ -13,7 +13,7 @@ mod constants;
 
 pub use constants::{BTRIT_0, BTRIT_1, BTRIT_NEG_1, UTRIT_0, UTRIT_1, UTRIT_2, UTRIT_U384_MAX, UTRIT_U384_MAX_HALF};
 
-use crate::hashes::sponge::kerl::bigint::{
+use crate::hashes::ternary::kerl::bigint::{
     binary_representation::{U32Repr, U8Repr},
     endianness::{BigEndian, LittleEndian},
     I384, T242, U384,
