+++ src/hashes/ternary/mod.rs
