+++ src/keys/ternary/seed.rs
@@ -3,7 +3,7 @@
 
 //! Ternary seed to derive private keys, public keys and signatures from.
 
-use crate::hashes::sponge::{kerl::Kerl, Sponge, HASH_LENGTH};
+use crate::hashes::ternary::{kerl::Kerl, Sponge, HASH_LENGTH};
 
 use bee_common_derive::{SecretDebug, SecretDisplay, SecretDrop};
 use bee_ternary::{Btrit, T1B1Buf, Trit, TritBuf, Trits, TryteBuf, T1B1};
